Move notebook diff
